move version and useragent definitions into sdk
